updated create_work_item to support markdown format in large text fields (#261)
Updated `create_work_item` to accept a `format` param so that you can
set large text fields to use markdown. By default it is HTML. Also note,
there is not good way on create to know if the field is a large text
field or not. So it is going to assume the LLM knows what you ared doing
based on your prompt. As a secondary check, we are looking at the value
and if it is greater than 50 characters, it is safer to assume it is a
large text field. Worse case, the opp will fail if you try and set a
field to markdown that is not large text.

Updated `update_work_items_batch` as well to take format

## GitHub issue number #218 #264

## **Associated Risks**

Should be fine, no risks

## ✅ **PR Checklist**

- [x] **I have read the [contribution
guidelines](https://github.com/microsoft/azure-devops-mcp/blob/main/CONTRIBUTING.md)**
- [x] **I have read the [code of conduct
guidelines](https://github.com/microsoft/azure-devops-mcp/blob/main/CODE_OF_CONDUCT.md)**
- [x] Title of the pull request is clear and informative.
- [x] 👌 Code hygiene
- [x] 🔭 Telemetry added, updated, or N/A
- [x] 📄 Documentation added, updated, or N/A
- [x] 🛡️ Automated tests added, or N/A

## 🧪 **How did you test it?**

updated tests and ran thru manual tests