Update README.md to clarify copilot settings (#388)
As described in #378 , the recommendation for copilot-instructions is
confusing and doesn't mention that this file is actually intended for
use in developing the ADO MCP server, but instead this repo is
recommending it for use _as_ the copilot instructions file even in other
repos. This recommendation seems wrong because other repos do not
necessarily contain `src/tools.ts` and related files - the only part
that is likely helpful for other repos is that it informs the AI to look
for ADO MCP tools in general. One of my teammates was confused by the
recommendation this week and was trying to merge this file into our
codebase which was confusing GitHub Copilot during our local testing.

To fix this, I am changing the recommendation in the README to instead
suggest putting "This project uses Azure DevOps. Always check to see if
the Azure DevOps MCP server has a tool relevant to the user's request"
in the copilot instructions file. Based on my local testing, I don't
think it is necessary to include other parts of the existing
copilot-instructions file, but it is still useful to have the AI look
for ADO tools by default instead of GitHub tools.

## GitHub issue number

#378

## **Associated Risks**

It is possible that the existing copilot instructions file is helping
someone whose codebase happens to specifically contain the same format
of tools at `src/tools.ts` and prompts at `src/prompts.ts`, and now new
users with the exact same specific codebase format will no longer gain
the benefit of the AI being able to immediately find those files.
However, I think this risk is extremely unlikely, and the AI can still
find those files.

## ✅ **PR Checklist**

- [x] **I have read the [contribution
guidelines](https://github.com/microsoft/azure-devops-mcp/blob/main/CONTRIBUTING.md)**
- [x] **I have read the [code of conduct
guidelines](https://github.com/microsoft/azure-devops-mcp/blob/main/CODE_OF_CONDUCT.md)**
- [x] Title of the pull request is clear and informative.
- [x] 👌 Code hygiene
- [x] 🔭 Telemetry added, updated, or N/A
- [x] 📄 Documentation added, updated, or N/A
- [x] 🛡️ Automated tests added, or N/A

## 🧪 **How did you test it?**

N/A readme change